use index 2 of attachments for name before sending
Phpmailer source code sets the filepath to index 0, the basename of the filepath to index 1, and an optional user specified name to index 2. If the user does not specify a custom name, it defers to the basename at index 1, but also sets it to index 2.

The Mailgun provider for this plugin should take the value at index 2 of attachments to use a possibly specified custom name that is defaulted to the basename.